requirements.txt
# Core Dependencies
requests>=2.25.1
beautifulsoup4>=4.9.3
selenium>=4.1.0
webdriver-manager>=3.5.2
playsound>=1.2.2
python-dotenv>=0.19.2
python-dateutil>=2.8.2
pytz>=2021.3

# Web Application Dependencies
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=3.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.0.0
Flask-Mail>=0.9.1
Werkzeug>=2.0.0

# Database Dependencies
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.7.0

# Payment Processing
stripe>=3.0.0

# Background Tasks
celery>=5.2.0
redis>=4.0.0

# Testing
pytest>=7.0.0
pytest-flask>=1.2.0
pytest-cov>=3.0.0

# Development
black>=22.0.0
flake8>=4.0.0
mypy>=0.950